+++ CHANGELOG.md
@@ -29,6 +29,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 - Allow Pushgateway to now require job names for compatibility with Gravel Gateway.
 - Allow `histogram.startTime()` to be used with exemplars.
+- added `defaultLabels` option to metric classes
 
 [15.1.0]: https://github.com/siimon/prom-client/compare/v15.0.0...v15.1.0